requirements.txt
annotated-types==0.7.0
anyio==4.7.0
click==8.1.7
fastapi==0.115.6
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
psycopg==3.2.3
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
python-multipart==0.0.19
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
